Move the extract.modify_save_file.get_element_lineage function to extract.extract_save_data.get_element_lineage
