+++ extract/extract_save_data.py
@@ -115,6 +115,44 @@ def get_elements_by_search_pattern(tree: xml.etree.ElementTree, element_search_p
     return element_list
 
 
+def get_element_lineage(element: xml.etree.ElementTree.Element,
+    root: xml.etree.ElementTree.Element, lineage: str=None) -> str:
+    """Return the lineage of an element as a string showing the hierarchy of tags going back to the
+    root element of the XML document
+
+    Parameters:
+    element (xml.etree.ElementTree.Element): The element being analyzed
+    root (xml.etree.ElementTree.Element): The root element of the XML document
+    lineage (str): The string representation of the element hierarchy
+
+    Returns:
+    str: String representation of the lineage of the XML element
+    """
+
+    # Add the starting element to the lineage string
+    if lineage is None:
+        logging.debug("Initializing the lineage string with the starting element")
+        lineage = element.tag
+
+    # Check to see if the element has a parent element
+    parent = root.find(f".//{element.tag}/..")
+
+    # The top level has been reached. End recursion by returning the lineage data
+    if parent is None:
+        logging.debug("Tag lineage analysis is complete. Returning lineage:\n%s", lineage)
+        return lineage
+
+    # This block only executes when there is a valid parent
+    logging.debug("The current element, %s, has a parent element, %s.\nRecursing XML tree",
+        element.tag, parent.tag)
+
+    # Prepend the detected parent to the lineage string
+    lineage = f"{parent.tag} > {lineage}"
+
+    # Recurse upward in the hierarchy, eventually returning the complete hierarchy
+    return get_element_lineage(element=parent, root=root, lineage=lineage)
+
+
 def get_save_file_data(save_file_path: str) -> xml.etree.ElementTree.Element:
     """Return the root element from the RimWorld save game XML data
